+++ src/fit.rs
@@ -1,5 +1,8 @@
+pub mod config;
 pub mod enorm;
+pub mod enums;
 mod fit;
 pub mod mpfit;
+pub mod status;
 
 pub use crate::fit::fit::{ImpedanceDataFitter, ModelParameter};
